requirements.txt
streamlit
plotly
yfinance
pandas
ta
coinbase-advanced-py
python-dotenv
pycoingecko
sqlalchemy
psycopg2-binary
git+https://git@github.com/bjahnke/blue_bird_pms.git